+++ ydb/core/persqueue/pqtablet/blob/ut/ya.make
@@ -1,5 +1,8 @@
 UNITTEST_FOR(ydb/core/persqueue/pqtablet/blob)
 
-SRCS(blob_ut.cpp)
+SRCS(
+    blob_ut.cpp
+    type_codecs_ut.cpp
+)
 
 END()
