+++ ydb/core/persqueue/ut/ya.make
@@ -42,7 +42,6 @@ SRCS(
     pqtablet_ut.cpp
     quota_tracker_ut.cpp
     sourceid_ut.cpp
-    type_codecs_ut.cpp
     user_info_ut.cpp
     pqrb_describes_ut.cpp
     partition_scale_manager_graph_cmp_ut.cpp
