+++ ydb/core/persqueue/ya.make
@@ -8,7 +8,6 @@ SRCS(
     list_all_topics_actor.cpp
     percentile_counter.cpp
     pq_l2_cache.cpp
-    type_codecs_defs.cpp
     write_meta.cpp
     dread_cache_service/caching_service.cpp
     write_id.cpp
